+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -47,6 +47,7 @@ MonoBehaviour:
   - EnemyMove
   - EnemySlot
   - EnemyCard
+  - EnemyUltimateCutIn
   DisableAutoOpenWizard: 1
   ShowSettings: 1
   DevRegionSetOnce: 0
