+++ Assets/Resources/TimeLine/eUltCatin.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 489897250c4f4284f94285960c4d8455
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
