+++ Assets/Scripts/Card/CardSlot.cs
@@ -1,3 +1,4 @@
+using Manager;
 using UniRx;
 using UnityEngine;
