+++ Assets/Scripts/Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b457a9f1b6c6734b8fbbe72106018f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
