+++ Assets/Scripts/Manager/AnimationManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8296a29f0b331334ab002487622567ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
